SCREAM: moved FieldRequest and GroupRequest to separate file
